# [GLUTEN-9197][CH] Simplify sum aggregate expression