--- a/backends-clickhouse/src/test/scala/org/apache/gluten/execution/GlutenFunctionValidateSuite.scala
+++ b/backends-clickhouse/src/test/scala/org/apache/gluten/execution/GlutenFunctionValidateSuite.scala
@@ -16,6 +16,8 @@
  */
 package org.apache.gluten.execution
 
+import org.apache.gluten.backendsapi.clickhouse.CHConfig
+
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.{DataFrame, GlutenTestUtils, Row}
 import org.apache.spark.sql.catalyst.expressions._
@@ -1045,7 +1047,7 @@ class GlutenFunctionValidateSuite extends GlutenClickHouseWholeStageTransformerS
     compareResultsAgainstVanillaSpark(sql, true, { _ => })
   }
 
-  test("GLUTEN-8723 fix slice unexpected exception") {
+  <<<<<<<.HEAD(test("GLUTEN-8723 fix slice unexpected exception") {
     val create_sql = "create table t_8723 (full_user_agent string) using orc"
     val insert_sql = "insert into t_8723 values(NULL)"
     val select1_sql = "select " +
@@ -1058,7 +1060,7 @@ class GlutenFunctionValidateSuite extends GlutenClickHouseWholeStageTransformerS
     compareResultsAgainstVanillaSpark(select1_sql, true, { _ => })
     compareResultsAgainstVanillaSpark(select2_sql, true, { _ => })
     spark.sql(drop_sql)
-  }
+  })
 
   test("GLUTEN-8715 nan semantics") {
     withTable("test_8715") {
@@ -1258,5 +1260,27 @@ class GlutenFunctionValidateSuite extends GlutenClickHouseWholeStageTransformerS
     // default comparator without array elements not nullable guaranteed
     val sql2 = "select array_sort(array(id+1, null, id+2)) from range(10)"
     compareResultsAgainstVanillaSpark(sql2, true, { _ => })
+  }
+
+  test("Test SimplifySumRule for sum simplification") {
+    val sql = "select sum(id / 3), sum(id * 7), sum(7 * id) from range(10)"
+
+    def checkSimplifiedSum(df: DataFrame): Unit = {
+      val projects = collectWithSubqueries(df.queryExecution.executedPlan) {
+        case project: ProjectExecTransformer
+            if project.child.isInstanceOf[CHHashAggregateExecTransformer] =>
+          project
+      }
+
+      assert(projects.size == 1)
+      assert(projects.head.projectList.size == 3)
+      assert(projects.head.projectList(0).asInstanceOf[Alias].child.isInstanceOf[Divide])
+      assert(projects.head.projectList(1).asInstanceOf[Alias].child.isInstanceOf[Multiply])
+      assert(projects.head.projectList(2).asInstanceOf[Alias].child.isInstanceOf[Multiply])
+    }
+
+    withSQLConf((CHConfig.runtimeConfig("enable_simplify_sum"), "true")) {
+      compareResultsAgainstVanillaSpark(sql, compareResult = true, checkSimplifiedSum)
+    }
   }
 }