## What changes were proposed in this pull request?

Simplify sum with the following rules: 
- Rule 1: sum(expr / literal) -> sum(expr) / literal
- Rule 2: sum(expr * literal) -> literal * sum(expr)
- Rule 3: sum(literal * expr) -> literal * sum(expr) 

If current rule is combined with https://github.com/apache/incubator-gluten/pull/9185, we can easily transform
- sum(if(cond, expr/literal, 0)) -> (sum(expr) filter (where cond)) / literal
- sum(if(cond, expr * literal, 0)) -> (sum(expr) filter (where cond)) * literal
- sum(if(cond, literal * expr, 0)) -> (sum(expr) filter (where cond)) * literal


(Fixes: \#9197)

## How was this patch tested?

(Please explain how this patch was tested. E.g. unit tests, integration tests, manual tests)


(If this patch involves UI changes, please attach a screenshot; otherwise, remove this)