@taiyang-li 2025-04-01
Performance tests

``` sql 
CREATE TEMPORARY VIEW lineitem
USING org.apache.spark.sql.parquet
OPTIONS (
  path  "/data1/liyang/tpch/tpch-data-sf1/lineitem"
) ; 

select 
sum(l_quantity * 1.234), 
sum(1.234 * l_tax), 
sum(l_extendedprice / 1.234)
from lineitem;  

After 
1 row selected (0.3 seconds)
1 row selected (0.292 seconds)
1 row selected (0.296 seconds)

Before
1 row selected (0.348 seconds)
1 row selected (0.355 seconds)
1 row selected (0.34 seconds)

```